requirements.txt
pandas
numpy
matplotlib
seaborn
requests
beautifulsoup4
scikit-learn
google-genai
fastapi
uvicorn[standard]
httpx
tweepy
python-dotenv
openai
orjson